WP-360 Using [desc] in listing details template truncates description at 500 chars
The description template tag was truncating at 500 characters on the Property Details Page (defined by listing details template) and for both search_results & static_listings shortcodes.  This was causing a loss of information on the PDP and it was also providing too much information for search_results and static_listings.

Modifications have been made so that text is not truncated on the Property Details page and the description field is being truncated at 15 words for both search_results and static_listings shortcodes.  A description of the 'maxlen' attribute was added to the description of the [desc] template tag which will allow users to optionally override the defaults.