WP-180
After speaking with Matt Williams it was determined that we do not need to offer the ability to toggle between exact/like matches for text type shortcode filters.  This change determines the type of filter being evaluated and if it's an agent id or an office id then an exact match will be performed otherwise a like match is performed.

It is not possible for the user to filter by more than one agent or office id from within the custom shortcode creation form.  It is possible for users to configure multiple agent or office ids within a filter by editing the long form shortcode but doing this will force them to explicitly declare the match type.  It will be set to a value of "in" so we do not need to account for that scenario.